Move column data building in api v3 to mix in
